Move forms related to people to people app
No functional changes, this is in an attempt to reduce the size of the
candidates model
