+++ ynr/apps/candidates/tests/test_validators.py
@@ -3,7 +3,7 @@ from unittest import skip
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from candidates.forms import BasePersonForm, UpdatePersonForm
+from people.forms import BasePersonForm, UpdatePersonForm
 
 from .factories import PersonFactory
 from .uk_examples import UK2015ExamplesMixin
